sway: Move wmenu_run to sway and use swaymsg exec
This commit changes the way the wmenu_run script works to use swaymsg to
execute the programs instead. This is the way wmenu is used in the
default config and it is my hope that it will fix some crashing problems
I've been having with sway where opening applications would sometimes
crash the entire compositor.

Since the script is now tightly coupled with sway, I decided to move it
to the "sway" module.