emacs: Remove package manager and switch to cobalt2
I'm bundling a cobalt2 theme with the config and then modifying it
to have green comments in my init.el, which is similar to a setup I
have in VS Code which I've been pretty happy with as of late.

Also removing the package management stuff because it just bloats up
my editor and I don't use it right now. Will MELPA back if it changes.