nvim: Replace some obsidian features
I want to ditch Obsidian for Neovim. So I'm just getting the subset of
features I need from Obsidian.

This commit adds plugins for:
1. Inline markdown rendering
2. Inline equation (mathjax) rendering
3. Image rendering

This, in addition to neowiki -- which implements the note links and
backlinks -- basically gets me Obsidian inside Neovim, at least the
parts of it that I use the most.